refactor(search): get public key from query headers
