+++ async-crypt4gh/Cargo.toml
@@ -28,4 +28,4 @@ bincode = "1.3.3"
 [dev-dependencies]
 noodles = { version = "0.60", features = ["async", "bam", "sam"] }
 
-htsget-test = { version = "0.5.0", path = "../htsget-test", features = ["http-tests"], default-features = false }
+htsget-test = { version = "0.5.0", path = "../htsget-test", features = ["http-tests", "crypt4gh"], default-features = false }
