+++ async-crypt4gh/src/lib.rs
@@ -226,35 +226,9 @@ impl Deref for DecryptedBytes {
 
 #[cfg(test)]
 pub(crate) mod tests {
-  use crypt4gh::keys::{get_private_key, get_public_key};
-  use crypt4gh::Keys;
   use tokio::io::AsyncReadExt;
 
-  use htsget_test::http_tests::{get_test_file, get_test_path};
-
-  /// Returns the private keys of the recipient and the senders public key from the context of decryption.
-  pub(crate) async fn get_decryption_keys() -> (Keys, Vec<u8>) {
-    get_keys("crypt4gh/keys/bob.sec", "crypt4gh/keys/alice.pub").await
-  }
-
-  /// Returns the private keys of the recipient and the senders public key from the context of encryption.
-  pub(crate) async fn get_encryption_keys() -> (Keys, Vec<u8>) {
-    get_keys("crypt4gh/keys/alice.sec", "crypt4gh/keys/bob.pub").await
-  }
-
-  async fn get_keys(private_key: &str, public_key: &str) -> (Keys, Vec<u8>) {
-    let private_key = get_private_key(get_test_path(private_key), Ok("".to_string())).unwrap();
-    let public_key = get_public_key(get_test_path(public_key)).unwrap();
-
-    (
-      Keys {
-        method: 0,
-        privkey: private_key,
-        recipient_pubkey: public_key.clone(),
-      },
-      public_key,
-    )
-  }
+  use htsget_test::http_tests::get_test_file;
 
   /// Get the original file bytes.
   pub(crate) async fn get_original_file() -> Vec<u8> {
